Not to set bit 8 in PML4E, PDPTE
This patch fixes crash in KVM-enabled AMD CPU.

As for Intel CPU, the 8th bit in PML4E is ignored.
But as for AMD CPU, the 8th bit in PML4E Must Be Zero.
Setting 8th bit in PML4E on AMD CPU causes crash.
We must not set the 8th bit in PML4E and PDPTE.

Reference
- Intel® 64 and IA-32 Architectures Software Developer’s Manual Vol.3, Figure 4-11
- AMD64 Architecture Programmer’s Manual Vol.3, Figure 5-20

Signed-off-by: Naoki MATSUMOTO <naoki@pibvt.net>